split & test queue
